backend/src/main/java/com/hiddenartist/backend/global/security/handler/OAuth2LogoutSuccessHandler.java
package com.hiddenartist.backend.global.security.handler;

import jakarta.servlet.ServletException;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import java.io.IOException;
import lombok.RequiredArgsConstructor;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.security.core.Authentication;
import org.springframework.security.web.authentication.logout.LogoutSuccessHandler;
import org.springframework.stereotype.Component;

@Component
@RequiredArgsConstructor
public class OAuth2LogoutSuccessHandler implements LogoutSuccessHandler {

  @Value("${redirect-url.logout}")
  private String REDIRECT_URL;

  @Override
  public void onLogoutSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication)
      throws IOException, ServletException {
    response.getWriter().write("로그아웃이 완료되었습니다. 홈페이지로 이동합니다.");
    response.sendRedirect(REDIRECT_URL);
  }

}